+++ db/models.py
@@ -3,9 +3,9 @@
 # pylint: disable=no-self-argument,no-name-in-module
 
 from datetime import datetime
-from typing import Dict, List, Literal, Optional
+from typing import Dict, List, Literal, Optional, Union
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, Field, validator
 
 
 class DocumentBaseModel(BaseModel):
@@ -38,7 +38,16 @@ class CircuitScheduleDoc(BaseModel):
 class TAGDoc(BaseModel):
     """TAG model"""
     tag_type: int
-    value: int
+    value: Union[int, str]
+
+    @validator('value')
+    def validate_value(cls, value):
+        """Validate value when is a string"""
+        if isinstance(value, int):
+            return value
+        if isinstance(value, str) and value in ("any", "untagged"):
+            return value
+        raise ValueError("value as string allows 'any' and 'untagged'")
 
 
 class UNIDoc(BaseModel):