+++ settings.py
@@ -41,6 +41,8 @@ BATCH_SIZE = 50
 # is not set in a request
 EVPL_SB_PRIORITY = 20000
 EPL_SB_PRIORITY = 10000
+ANY_SB_PRIORITY = 15000
+UNTAGGED_SB_PRIORITY = 20000
 
 #  Time (seconds) to check if an evc has been updated
 #  or flows have been deleted.