+++ tests/unit/models/test_evc_deploy.py
@@ -18,7 +18,9 @@ from napps.kytos.mef_eline.settings import (
     MANAGER_URL,
     SDN_TRACE_CP_URL,
     EVPL_SB_PRIORITY,
-    EPL_SB_PRIORITY
+    EPL_SB_PRIORITY,
+    ANY_SB_PRIORITY,
+    UNTAGGED_SB_PRIORITY
 )  # NOQA
 from napps.kytos.mef_eline.exceptions import FlowModException  # NOQA
 from napps.kytos.mef_eline.tests.helpers import (
@@ -210,7 +212,8 @@ class TestEVC(TestCase):
         )
 
         expected_flow_mod = {
-            "match": {"in_port": interface_a.port_number, "dl_vlan": in_vlan},
+            "match": {"in_port": interface_a.port_number,
+                      "dl_vlan": in_vlan},
             "cookie": evc.get_cookie(),
             "actions": [
                 {"action_type": "pop_vlan"},
@@ -220,6 +223,7 @@ class TestEVC(TestCase):
         }
         self.assertEqual(expected_flow_mod, flow_mod)
 
+    # pylint: disable=too-many-branches
     def test_prepare_push_flow(self):
         """Test prepare push flow method."""
         attributes = {
@@ -233,8 +237,8 @@ class TestEVC(TestCase):
         interface_z = evc.uni_z.interface
         out_vlan_a = 20
 
-        for in_vlan_a in (10, None):
-            for in_vlan_z in (3, None):
+        for in_vlan_a in [100, "4096/4096", 0, None]:
+            for in_vlan_z in [50, "4096/4096", 0, None]:
                 with self.subTest(in_vlan_a=in_vlan_a, in_vlan_z=in_vlan_z):
                     # pylint: disable=protected-access
                     flow_mod = evc._prepare_push_flow(interface_a, interface_z,
@@ -251,28 +255,30 @@ class TestEVC(TestCase):
                                 'port': interface_z.port_number
                             }
                         ],
-                        "priority": EPL_SB_PRIORITY,
+                        "priority": EVPL_SB_PRIORITY,
                     }
-                    if in_vlan_a and in_vlan_z:
+                    expected_flow_mod["priority"] = evc.get_priority(in_vlan_a)
+                    if in_vlan_a is not None:
                         expected_flow_mod['match']['dl_vlan'] = in_vlan_a
-                        expected_flow_mod['actions'].insert(0, {
-                            'action_type': 'set_vlan', 'vlan_id': in_vlan_z
-                        })
-                        expected_flow_mod['priority'] = EVPL_SB_PRIORITY
-                    elif in_vlan_a:
-                        expected_flow_mod['match']['dl_vlan'] = in_vlan_a
-                        expected_flow_mod['actions'].insert(0, {
-                            'action_type': 'pop_vlan'
-                        })
-                        expected_flow_mod["priority"] = EVPL_SB_PRIORITY
-                    elif in_vlan_z:
-                        expected_flow_mod['actions'].insert(0, {
-                            'action_type': 'set_vlan', 'vlan_id': in_vlan_z
-                        })
-                        expected_flow_mod['actions'].insert(0, {
-                            'action_type': 'push_vlan', 'tag_type': 'c'
-                        })
-                        expected_flow_mod["priority"] = EPL_SB_PRIORITY
+
+                    if in_vlan_z not in evc.special_cases:
+                        new_action = {"action_type": "set_vlan",
+                                      "vlan_id": in_vlan_z}
+                        expected_flow_mod["actions"].insert(0, new_action)
+
+                    if in_vlan_a not in evc.special_cases:
+                        if in_vlan_z == 0:
+                            new_action = {"action_type": "pop_vlan"}
+                            expected_flow_mod["actions"].insert(0, new_action)
+                    elif in_vlan_a == "4096/4096":
+                        if in_vlan_z == 0:
+                            new_action = {"action_type": "pop_vlan"}
+                            expected_flow_mod["actions"].insert(0, new_action)
+                    elif not in_vlan_a:
+                        if in_vlan_z not in evc.special_cases:
+                            new_action = {"action_type": "push_vlan",
+                                          "tag_type": "c"}
+                            expected_flow_mod["actions"].insert(0, new_action)
                     self.assertEqual(expected_flow_mod, flow_mod)
 
     @staticmethod
@@ -467,7 +473,7 @@ class TestEVC(TestCase):
                     {"action_type": "set_vlan", "vlan_id": in_vlan},
                     {"action_type": "output", "port": in_port},
                 ],
-                "priority": EVPL_SB_PRIORITY
+                "priority": EVPL_SB_PRIORITY,
             },
         ]
 
@@ -1102,131 +1108,87 @@ class TestEVC(TestCase):
         }
         return EVC(**attributes)
 
-    @staticmethod
+    # pylint: disable=too-many-branches
     @patch("napps.kytos.mef_eline.models.evc.EVC._send_flow_mods")
-    def test_deploy_direct_uni_flows(send_flow_mods_mock):
-        """Test _install_direct_uni_flows."""
+    def test_deploy_direct_uni_flows(self, send_flow_mods_mock):
+        """Test _install_direct_uni_flows"""
         evc = TestEVC.create_evc_intra_switch()
-
-        # Test 1: both UNIs with TAG
         expected_dpid = evc.uni_a.interface.switch.id
-        expected_flows = [
-            {
-                "match": {"in_port": 1, "dl_vlan": 82},
-                "cookie": evc.get_cookie(),
-                "actions": [
-                    {"action_type": "set_vlan", "vlan_id": 84},
-                    {"action_type": "output", "port": 3},
-                ],
-                "priority": EVPL_SB_PRIORITY
-            },
-            {
-                "match": {"in_port": 3, "dl_vlan": 84},
-                "cookie": evc.get_cookie(),
-                "actions": [
-                    {"action_type": "set_vlan", "vlan_id": 82},
-                    {"action_type": "output", "port": 1},
-                ],
-                "priority": EVPL_SB_PRIORITY
-            }
-        ]
 
-        # pylint: disable=protected-access
-        evc._install_direct_uni_flows()
-        send_flow_mods_mock.assert_called_once_with(
-            expected_dpid, expected_flows
-        )
+        for uni_a in [100, "4096/4096", 0, None]:
+            for uni_z in [100, "4096/4096", 0, None]:
+                with self.subTest(uni_a=uni_a, uni_z=uni_z):
+                    expected_flows = [
+                        {
+                            "match": {"in_port": 1},
+                            "cookie": evc.get_cookie(),
+                            "actions": [
+                                {"action_type": "output", "port": 3},
+                            ],
+                            "priority": EPL_SB_PRIORITY
+                        },
+                        {
+                            "match": {"in_port": 3},
+                            "cookie": evc.get_cookie(),
+                            "actions": [
+                                {"action_type": "output", "port": 1},
+                            ],
+                            "priority": EPL_SB_PRIORITY
+                        }
+                    ]
+                    evc.uni_a = get_uni_mocked(tag_value=uni_a, is_valid=True)
+                    evc.uni_a.interface.port_number = 1
+                    evc.uni_z = get_uni_mocked(tag_value=uni_z, is_valid=True)
+                    evc.uni_z.interface.port_number = 3
+                    expected_dpid = evc.uni_a.interface.switch.id
+                    evc._install_direct_uni_flows()
+                    if uni_a is not None:
+                        expected_flows[0]["match"]["dl_vlan"] = uni_a
+                    if uni_z is not None:
+                        expected_flows[1]["match"]["dl_vlan"] = uni_z
+                    expected_flows[0]["priority"] = EVC.get_priority(uni_a)
+                    expected_flows[1]["priority"] = EVC.get_priority(uni_z)
 
-        # Test2: no TAG in UNI_A
-        uni_a_tag = evc.uni_a.user_tag
-        evc.uni_a.user_tag = None
-        expected_flows_no_tag_a = [
-            {
-                "match": {"in_port": 1},
-                "cookie": evc.get_cookie(),
-                "actions": [
-                    {"action_type": "set_vlan", "vlan_id": 84},
-                    {"action_type": "output", "port": 3},
-                ],
-                "priority": EPL_SB_PRIORITY
-            },
-            {
-                "match": {"in_port": 3, "dl_vlan": 84},
-                "cookie": evc.get_cookie(),
-                "actions": [
-                    {"action_type": "pop_vlan"},
-                    {"action_type": "output", "port": 1},
-                ],
-                "priority": EVPL_SB_PRIORITY
-            }
-        ]
-        evc._install_direct_uni_flows()
-        send_flow_mods_mock.assert_called_with(
-            expected_dpid, expected_flows_no_tag_a
-        )
-        evc.uni_a.user_tag = uni_a_tag
+                    if uni_z not in evc.special_cases:
+                        expected_flows[0]["actions"].insert(
+                            0, {"action_type": "set_vlan", "vlan_id": uni_z}
+                        )
 
-        # Test3: no TAG in UNI_Z
-        uni_z_tag = evc.uni_z.user_tag
-        evc.uni_z.user_tag = None
-        expected_flows_no_tag_z = [
-            {
-                "match": {"in_port": 1, "dl_vlan": 82},
-                "cookie": evc.get_cookie(),
-                "actions": [
-                    {"action_type": "pop_vlan"},
-                    {"action_type": "output", "port": 3},
-                ],
-                "priority": EVPL_SB_PRIORITY
-            },
-            {
-                "match": {"in_port": 3},
-                "cookie": evc.get_cookie(),
-                "actions": [
-                    {"action_type": "set_vlan", "vlan_id": 82},
-                    {"action_type": "output", "port": 1},
-                ],
-                "priority": EPL_SB_PRIORITY
-            }
-        ]
-        evc._install_direct_uni_flows()
-        send_flow_mods_mock.assert_called_with(
-            expected_dpid, expected_flows_no_tag_z
-        )
-        evc.uni_z.user_tag = uni_z_tag
-
-        # Test3: no TAG in both UNI_Z and UNI_Z
-        evc.uni_a.user_tag = None
-        evc.uni_z.user_tag = None
-        expected_flows_no_tag = [
-            {
-                "match": {"in_port": 1},
-                "cookie": evc.get_cookie(),
-                "actions": [{"action_type": "output", "port": 3}],
-                "priority": EPL_SB_PRIORITY
-            },
-            {
-                "match": {"in_port": 3},
-                "cookie": evc.get_cookie(),
-                "actions": [{"action_type": "output", "port": 1}],
-                "priority": EPL_SB_PRIORITY
-            }
-        ]
-        evc._install_direct_uni_flows()
-        send_flow_mods_mock.assert_called_with(
-            expected_dpid, expected_flows_no_tag
-        )
-#
-#        print(evc._prepare_direct_uni_flows())
-#        evc.uni_a.user_tag = uni_a_tag
-#        uni_z_tag = evc.uni_z.user_tag
-#        evc.uni_z.user_tag = None
-#        print(evc._prepare_direct_uni_flows())
-#        evc.uni_z.user_tag = uni_z_tag
-#        evc.uni_a.user_tag = None
-#        evc.uni_z.user_tag = None
-#        print(evc._prepare_direct_uni_flows())
-#        self.assertTrue(False)
+                    if uni_a not in evc.special_cases:
+                        expected_flows[1]["actions"].insert(
+                                0, {"action_type": "set_vlan",
+                                    "vlan_id": uni_a}
+                            )
+                        if not uni_z:
+                            expected_flows[1]["actions"].insert(
+                                0, {"action_type": "push_vlan",
+                                    "tag_type": "c"}
+                            )
+                        if uni_z == 0:
+                            new_action = {"action_type": "pop_vlan"}
+                            expected_flows[0]["actions"].insert(0, new_action)
+                    elif uni_a == "4096/4096":
+                        if uni_z == 0:
+                            new_action = {"action_type": "pop_vlan"}
+                            expected_flows[0]["actions"].insert(0, new_action)
+                    elif uni_a == 0:
+                        if uni_z not in evc.special_cases:
+                            expected_flows[0]["actions"].insert(
+                                0, {"action_type": "push_vlan",
+                                    "tag_type": "c"}
+                            )
+                        if uni_z:
+                            new_action = {"action_type": "pop_vlan"}
+                            expected_flows[1]["actions"].insert(0, new_action)
+                    elif uni_a is None:
+                        if uni_z not in evc.special_cases:
+                            expected_flows[0]["actions"].insert(
+                                0, {"action_type": "push_vlan",
+                                    "tag_type": "c"}
+                            )
+                    send_flow_mods_mock.assert_called_with(
+                        expected_dpid, expected_flows
+                    )
 
     def test_is_affected_by_link(self):
         """Test is_affected_by_link method"""
@@ -1639,3 +1601,35 @@ class TestEVC(TestCase):
         self.evc_deploy.primary_path = Path([])
         self.evc_deploy.backup_path = Path([])
         self.assertTrue(self.evc_deploy.is_eligible_for_failover_path())
+
+    def test_get_value_from_uni_tag(self):
+        """Test _get_value_from_uni_tag"""
+        uni = get_uni_mocked(tag_value=None)
+        value = EVC._get_value_from_uni_tag(uni)
+        self.assertEqual(value, None)
+
+        uni = get_uni_mocked(tag_value="any")
+        value = EVC._get_value_from_uni_tag(uni)
+        self.assertEqual(value, "4096/4096")
+
+        uni = get_uni_mocked(tag_value="untagged")
+        value = EVC._get_value_from_uni_tag(uni)
+        self.assertEqual(value, 0)
+
+        uni = get_uni_mocked(tag_value=100)
+        value = EVC._get_value_from_uni_tag(uni)
+        self.assertEqual(value, 100)
+
+    def test_get_priority(self):
+        """Test get_priority_from_vlan"""
+        evpl_value = EVC.get_priority(100)
+        self.assertEqual(evpl_value, EVPL_SB_PRIORITY)
+
+        untagged_value = EVC.get_priority(0)
+        self.assertEqual(untagged_value, UNTAGGED_SB_PRIORITY)
+
+        any_value = EVC.get_priority("4096/4096")
+        self.assertEqual(any_value, ANY_SB_PRIORITY)
+
+        epl_value = EVC.get_priority(None)
+        self.assertEqual(epl_value, EPL_SB_PRIORITY)